fix: `ModuleNotFoundError` when using telemetry in Python client
